requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
requests>=2.25.0
pillow>=8.0.0
scikit-learn>=1.0.0
seaborn>=0.11.0